requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.1.0
google-generativeai>=0.8.0
pydantic>=2.10.0
python-dotenv>=1.1.0
sse-starlette>=2.2.0
